+++ src/main/kotlin/uk/co/ben_gibson/git/link/ui/Icons.kt
@@ -11,7 +11,6 @@ object Icons {
     val GITEA = IconLoader.getIcon("/icons/gitea.svg", javaClass)
     val GIT = IconLoader.getIcon("/icons/git.svg", javaClass)
     val GIT_LINK = IconLoader.getIcon("/icons/gitlink.svg", javaClass)
-    val CHROMIUM = IconLoader.getIcon("/icons/chromium.svg", javaClass)
     val GITEE = IconLoader.getIcon("/icons/gitee.svg", javaClass)
     val GERRIT = IconLoader.getIcon("/icons/gerrit.svg", javaClass)
     val CODEBERG = IconLoader.getIcon("/icons/codeberg.svg", javaClass)